changes made in controllers and listings folders

+++ controllers/reviews.js
@@ -3,6 +3,11 @@ const Listing = require("../models/listing");
 
 module.exports.createReview = async (req, res) => {
   let listing = await Listing.findById(req.params.id);
+  if(!listing){
+    req.flash("error","Listing not found!");
+    return res.redirect("/listings");
+  }
+  
   let newReview = new Review(req.body.review);
   newReview.author = req.user._id;
 
@@ -10,13 +15,18 @@ module.exports.createReview = async (req, res) => {
   await newReview.save();
   await listing.save();
   req.flash("success", "New review added!");
-  res.redirect(`/listings/${listing._id}`);
+  return res.redirect(`/listings/${listing._id}`);
 };
 
 module.exports.destroyReview = async (req, res) => {
   let { id, reviewId } = req.params;
+  let listing = await Listing.findById(id);
+  if(!listing){
+    req.flash("error", "Listing not found!");
+    return res.redirect("/listings");
+  }
   await Listing.findByIdAndUpdate(id, { $pull: { reviews: reviewId } });
   await Review.findByIdAndDelete(reviewId);
   req.flash("success", "Review deleted!");
-  res.redirect(`/listings/${id}`);
+  return res.redirect(`/listings/${id}`);
 };
